+++ apps/remix-ide/src/app/plugins/ViewPluginUI.tsx
@@ -1,0 +1,29 @@
+import React from "react"
+import { useEffect, useState } from "react"
+
+interface IViewPluginUI {
+    plugin: any
+  }
+  
+export const ViewPluginUI = (props: IViewPluginUI) => {
+  
+    const [state, setState] = useState<any>(null)
+  
+    useEffect(() => {
+      console.log(props.plugin)
+      if(props.plugin.setDispatch){
+        props.plugin.setDispatch(setState)
+      }
+    }, [])
+  
+    useEffect(() => {
+      console.log(state)
+    }, [state])
+  
+    return (
+      <>{state? 
+        <div>{props.plugin.updateComponent(state)}</div>
+      :<></>
+      }</>
+    )
+  }
